tests/kafka-test/requirements.txt
kafka-python>=2.0.2
pytest>=7.0.0
